+++ src/index.ts
@@ -1,5 +1,5 @@
-// Import CSS variables to bundle them
-import '../styles/src/variables.css'
+// Import CSS bundle for styles
+import './react/styles-bundle.css'
 
 // Export all React components (default export)
 export * from './react'